+++ types/my-types.d.ts
@@ -7,6 +7,7 @@ import { handleBonusesFor } from '../src/target-and-bonus-join.mjs';
 import { showBonusPicker } from '../src/handlebars-handlers/bonus-picker.mjs';
 import { BaseTargetOverride } from '../src/targeted/target-overides/_base-target-override.mjs';
 import { simplifyRollFormula } from '../src/util/simplify-roll-formula.mjs';
+import { VitalStrikeData } from '../src/bonuses/vital-strike.mjs';
 
 export {};
 
@@ -103,6 +104,7 @@ declare global {
             handleBonusesFor: typeof handleBonusesFor;
             array: Record<string, (...args) => any>;
             simplifyRollFormula: typeof simplifyRollFormula;
+            VitalStrikeData: typeof VitalStrikeData;
             [key: string]: any;
         };
     }